コードをすっきりさせた
